pyproject.toml
[tool.poetry]
name = "tuting"
version = "0.1.0"
description = "Accounting for every tutor"
authors = ["Art Hab <clarkreiz@gmail.com>"]
packages = [
    { include = "*", from = "src" }
]

[tool.poetry.scripts]
tuting = "tuting_bot.main:run_main"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.12.0"
pytest-asyncio = "^0.24.0"
pytest-aiohttp = "^1.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.2"
pytest-cov = "5.0.0"
mypy = "1.11.1"
bandit = "1.7.9"
docformatter = "1.7.5"
ruff = "0.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
backend-path = ["src"]

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = "tests"

[tool.ruff]
extend-exclude = [
    "__pycache__",
    "build",
    "dist",
]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
extend-select = [
    "C4",
    "D200",
    "D201",
    "D204",
    "D205",
    "D206",
    "D210",
    "D211",
    "D213",
    "D300",
    "D400",
    "D402",
    "D403",
    "D404",
    "D419",
    "E",
    "F",
    "G010",
    "I001",
    "INP001",
    "N805",
    "PERF101",
    "PERF102",
    "PERF401",
    "PERF402",
    "PGH004",
    "PGH005",
    "PIE794",
    "PIE796",
    "PIE807",
    "PIE810",
    "RET502",
    "RET503",
    "RET504",
    "RET505",
    "RUF015",
    "RUF100",
    "S101",
    "T20",
    "W",
]


[tool.mypy]
files = ["src", "tests"]
strict = "true"

[lint.per-file-ignores]
"tests/*.py" = ["S101"]
